+++ dom/DomBuilder.js
@@ -2,9 +2,11 @@
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.DomBuilder = void 0;
 /** A wrapper for a DOM element (similar to how a JQuery object wraps one or more DOM elements).
- * Reduces the code required to create and setup a new element for insertion into a Document.
- * Includes functions to set the element's ID, add class names, add styles, set attributes, set the element's text content, and add child elements.
- * Access the 'element' property at any time to retrieve the underlying DOM element.
+ * Exposes a builder pattern to reduce the code required to create and setup a new element for
+ * insertion into a Document.
+ * Includes functions to set the element's ID, add class names, add styles, set attributes,
+ * set the element's text content, and add child elements.
+ * Access the `element` property at any time to retrieve the underlying DOM element.
  * @author TeamworkGuy2
  * @since 2016-04-26
  */