+++ dom/DomBuilderHelper.ts
@@ -1,8 +1,9 @@
 ﻿import { BuilderHelper } from "dom-builder";
 
 /** Helper functions for XMLDocument Node attributes and children.
- * Includes a global {@link getParser} and {@link getSerializer} which default to {@link DOMParser} and {@link XMLSerializer},
- * but can be overridden.
+ * Useful for manipulating existing elements.
+ * Includes a global {@link getParser} and {@link getSerializer} which default to {@link DOMParser}
+ * and {@link XMLSerializer}, and can be overridden with {@link setParser} and {@link setSerializer}.
  * @since 2016-04-27
  */
 export class DomBuilderHelper implements BuilderHelper {